Unifies duplicate regex constants to avoid mismatches.
