+++ pkg/flow/ent/schema/ref.go
@@ -9,6 +9,7 @@ import (
 	"entgo.io/ent/schema/edge"
 	"entgo.io/ent/schema/field"
 	"entgo.io/ent/schema/index"
+	"github.com/direktiv/direktiv/pkg/util"
 	"github.com/google/uuid"
 )
 
@@ -22,7 +23,7 @@ func (Ref) Fields() []ent.Field {
 	return []ent.Field{
 		field.UUID("id", uuid.UUID{}).Default(uuid.New).Immutable().StorageKey("oid").StructTag(`json:"-"`),
 		field.Bool("immutable").Default(true).Immutable(),
-		field.String("name").Match(RefRegex).Immutable().Annotations(entgql.OrderField("NAME")),
+		field.String("name").Match(util.RefRegex).Immutable().Annotations(entgql.OrderField("NAME")),
 		field.Time("created_at").Default(time.Now).Immutable().Annotations(entgql.OrderField("CREATED")),
 	}
 }
