+++ pkg/model/function-common.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/direktiv/direktiv/pkg/flow/ent/schema"
+	"github.com/direktiv/direktiv/pkg/util"
 )
 
 // -------- Function Types -----------
@@ -118,7 +118,7 @@ func (o FunctionFileDefinition) Validate() error {
 		return errors.New("key required")
 	}
 
-	if !schema.VarNameRegex.MatchString(o.Key) {
+	if !util.VarNameRegex.MatchString(o.Key) {
 		return fmt.Errorf("key is invalid: must start with a letter and only contain letters, numbers and '_'")
 	}
