+++ pkg/model/state-getter.go
@@ -4,7 +4,6 @@ import (
 	"errors"
 	"fmt"
 
-	"github.com/direktiv/direktiv/pkg/flow/ent/schema"
 	"github.com/direktiv/direktiv/pkg/util"
 )
 
@@ -40,7 +39,7 @@ func (o *GetterDefinition) Validate() error {
 		return errors.New(`key required`)
 	}
 
-	if !schema.VarNameRegex.MatchString(o.Key) {
+	if !util.VarNameRegex.MatchString(o.Key) {
 		return fmt.Errorf("key is invalid: must start with a letter and only contain letters, numbers and '_'")
 	}
