+++ pkg/util/regex.go
@@ -5,8 +5,28 @@ import (
 	"regexp"
 )
 
-const RegexPattern = `^(([a-z][a-z0-9_\-]*[a-z0-9])|([a-z]))$`
-const VarRegexPattern = `^(([a-zA-Z][a-zA-Z0-9_\-]*[a-zA-Z0-9])|([a-zA-Z]))$`
+const NameRegexFragment = `(([a-z][a-z0-9_\-\.]*[a-z0-9])|([a-z]))`
+
+const NameRegexPattern = `^` + NameRegexFragment + `$`
+
+var NameRegex = regexp.MustCompile(NameRegexPattern)
+
+const URIRegexPattern = `^(` + NameRegexFragment + `[\/]?)*$`
+
+var URIRegex = regexp.MustCompile(URIRegexPattern)
+
+const VarNameRegexPattern = `^(([a-zA-Z][a-zA-Z0-9_\-\.]*[a-zA-Z0-9])|([a-zA-Z]))$`
+
+var VarNameRegex = regexp.MustCompile(VarNameRegexPattern)
+
+const RefRegexFragment = `(([a-zA-Z0-9][a-zA-Z0-9_\-\.]*[a-zA-Z0-9])|([a-zA-Z0-9]))`
+
+const RefRegexPattern = `^` + RefRegexFragment + `$`
+
+var RefRegex = regexp.MustCompile(RefRegexPattern)
+
+const RegexPattern = NameRegexPattern
+const VarRegexPattern = VarNameRegexPattern
 
 var reg *regexp.Regexp
 var varreg *regexp.Regexp
